Re-biased Fuel by 1/10, as the fuel required to reach the end was numerically dominating the other resources.
git-svn-id: file:///home/steve/svn/G3@192 2c4f788c-be06-4e6c-9a5a-78778216f3cc

+++ planets.py
@@ -47,7 +47,7 @@ class Planet():
         self.mining_hit_counter = MINING_HIT_COUNTER
         self.mining_can = {}
         for resource in const.initial_planetary_resources:
-            amount = int(const.initial_planetary_resources[resource] *  my_random.my_random())
+            amount = const.initial_planetary_resources[resource] *  my_random.my_random()
             if amount:
                 self.resources[resource] = amount
